WIP patch legacy routes
