+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/ScrambleHandler.kt
@@ -15,6 +15,8 @@ import org.worldcubeassociation.tnoodle.server.RouteHandler.Companion.splitNameA
 import org.worldcubeassociation.tnoodle.server.util.ServerEnvironmentConfig
 import org.worldcubeassociation.tnoodle.server.webscrambles.InvalidScrambleRequestException
 import org.worldcubeassociation.tnoodle.server.webscrambles.ScrambleRequest
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFBuilder
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFBindingGenerator
 import java.time.LocalDateTime
 
 class ScrambleHandler(val environmentConfig: ServerEnvironmentConfig) : RouteHandler {
@@ -51,6 +53,8 @@ class ScrambleHandler(val environmentConfig: ServerEnvironmentConfig) : RouteHan
                 ScrambleRequest.parseScrambleRequest(title, reqUrl, seed)
             }
 
+            val wcif = WCIFBindingGenerator.requestsToPseudoWCIF(scrambleRequests)
+
             when (extension) {
                 "txt" -> {
                     val sb = StringBuilder()
@@ -72,7 +76,7 @@ class ScrambleHandler(val environmentConfig: ServerEnvironmentConfig) : RouteHan
                 }
                 "json" -> call.respond(scrambleRequests)
                 "pdf" -> {
-                    val totalPdfOutput = ScrambleRequest.requestsToCompletePdf(title, generationDate.toLocalDate(), environmentConfig.projectTitle, scrambleRequests)
+                    val totalPdfOutput = WCIFBuilder.wcifToCompletePdf(wcif, generationDate.toLocalDate(), environmentConfig.projectTitle)
                     call.response.header("Content-Disposition", "inline")
 
                     // Workaround for Chrome bug with saving PDFs:
@@ -82,7 +86,8 @@ class ScrambleHandler(val environmentConfig: ServerEnvironmentConfig) : RouteHan
                     call.respondBytes(totalPdfOutput.render(), ContentType.Application.Pdf)
                 }
                 "zip" -> {
-                    val baosZip = ScrambleRequest.requestsToZip(title, generationDate, environmentConfig.projectTitle, scrambleRequests, seed, generationUrl, null)
+                    val modelZip = WCIFBuilder.wcifToZip(wcif, null, generationDate, environmentConfig.projectTitle, generationUrl.orEmpty())
+                    val baosZip = modelZip.compress()
 
                     call.respondBytes(baosZip, ContentType.Application.Zip)
                 }
