+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/routing/WcifHandler.kt
@@ -7,8 +7,8 @@ import io.ktor.routing.Routing
 import io.ktor.routing.post
 import org.worldcubeassociation.tnoodle.server.RouteHandler
 import org.worldcubeassociation.tnoodle.server.RouteHandler.Companion.parseQuery
+import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFBindingGenerator
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFParser
-import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFRequestBinding.Companion.generateBindings
 
 object WcifHandler : RouteHandler {
     override fun install(router: Routing) {
@@ -20,16 +20,9 @@ object WcifHandler : RouteHandler {
                 ?: return@post call.respond("Please specify a WCIF JSON")
 
             val wcif = WCIFParser.parseComplete(wcifJson)
+            val bindings = WCIFBindingGenerator.fillScrambleSets(wcif)
 
-            val title = query["title"]
-                ?: return@post call.respond("Please specify a title for the request set")
-
-            val bindings = wcif.generateBindings(title)
-
-            val flatRequests = bindings.activityScrambleRequests
-                .values.flatten()
-
-            call.respond(flatRequests)
+            call.respond(bindings)
         }
     }
 }
